test(modal): split up card tests
