+++ core/src/components/modal/test/card/modal-card.e2e.ts-snapshots/modal-card-custom-present-ios-ltr-Mobile-Firefox-linux.png
