+++ core/src/components/modal/test/card/modal-card.e2e.ts-snapshots/modal-card-present-ios-rtl-Mobile-Safari-linux.png
