+++ core/src/components/modal/test/card/modal-tablet.e2e.ts-snapshots/modal-card-custom-present-tablet-ios-ltr-Mobile-Firefox-linux.png
