+++ core/src/components/modal/test/card/modal-tablet.e2e.ts-snapshots/modal-card-present-tablet-ios-ltr-Mobile-Chrome-linux.png
